Addition of a script: profiles->workload.
An example input is also given. If some profiles are described in a JSON file,
Batsim workloads can be generated from those profiles thanks to this script.

To do so, I chose to:
    - generate N jobs
    - modelize interarrival times of jobs by a Weibull distribution
    - modelize job sized by a lognormal distribution

The distributions were chosen from the Chapter 9 of the book
"Workload modeling for performance evaluation" authored by Dror Feitelson.
The associated parameters can be given to this script as parameters,
try running it with --help to display all available options.